requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
psycopg2-binary~=2.9.10
pydantic~=2.11.4
python-jose~=3.4.0
passlib~=1.7.4
starlette~=0.46.2